Moved to flatter drivers modules
