+++ tests/unit/drivers/model_tests.py
@@ -1,7 +1,7 @@
 """SQLAlchemy model driver module."""
 from datetime import date
 
-from jsonapi_query.translation.model.sqlalchemy import SQLAlchemyDriver
+from jsonapi_query.drivers.model import SQLAlchemyDriver
 from tests.sqlalchemy import (
     BaseSQLAlchemyTestCase, Person, School, Student, Category)
