+++ tests/unit/drivers/view_tests.py
@@ -1,6 +1,6 @@
 from datetime import date
 
-from jsonapi_query.translation.view.marshmallow_jsonapi import MarshmallowDriver
+from jsonapi_query.drivers.view import MarshmallowDriver
 from tests.marshmallow_jsonapi import *
